refactor: extract StringArray component
