+++ app/components/FieldString/FieldStringEdit.jsx
@@ -1,23 +1,13 @@
 import Button from 'components/Button/Button'
+import formatLink from 'lib/util/formatLink'
 import Label from 'components/Label/Label'
 import proptypes from 'prop-types'
 import React from 'react'
 import RichEditor from 'components/RichEditor/RichEditor'
+import StringArray from 'components/StringArray/StringArray'
 import Style from 'lib/Style'
 import styles from './FieldString.css'
 import TextInput from 'components/TextInput/TextInput'
-
-function getLink(value, displayDirective) {
-  if (displayDirective && value) {
-    if (typeof displayDirective === 'string') {
-      return displayDirective.replace('{value}', value)
-    }
-
-    return /^https?:\/\//.test(value) ? value : 'http://' + value
-  }
-
-  return null
-}
 
 /**
  * Component for API fields of type String
@@ -126,73 +116,10 @@ export default class FieldStringEdit extends React.Component {
     super(props)
 
     this.cursor = null
-    this.dragBottomLimit = null
-    this.dragStart = {}
     this.inputRefs = []
-    this.itemRefs = []
-    this.listContainerRef = null
-
-    this.mousePositionListener = ({clientY: currentY}) => {
-      this.setState({currentY})
-
-      const {dragCurrentIndex: curr, dragOriginalIndex: orig} = this.state
-
-      if (curr > 0) {
-        const prevIndex = curr > orig ? curr : curr - 1
-        const prevItem = this.itemRefs[prevIndex].getBoundingClientRect()
-
-        if (prevItem.bottom - 5 > currentY) {
-          this.setState({
-            dragCurrentIndex: curr - 1
-          })
-        }
-      }
-
-      if (curr < this.props.value.length - 1) {
-        const nextIndex = curr < orig ? curr : curr + 1
-        const nextItem = this.itemRefs[nextIndex].getBoundingClientRect()
-
-        if (nextItem.top + 5 < currentY) {
-          this.setState({
-            dragCurrentIndex: curr + 1
-          })
-        }
-      }
-    }
-
-    this.dragEndListener = () => {
-      // prettier-ignore
-      window.document.removeEventListener('mousemove', this.mousePositionListener)
-      window.document.removeEventListener('mouseup', this.dragEndListener)
-
-      const {dragCurrentIndex, dragOriginalIndex} = this.state
-      const {value: valueArray, onChange} = this.props
-
-      if (dragCurrentIndex !== dragOriginalIndex) {
-        const newValueArray = valueArray.slice()
-        const [movedItem] = newValueArray.splice(dragOriginalIndex, 1)
-
-        newValueArray.splice(dragCurrentIndex, 0, movedItem)
-
-        typeof onChange === 'function' && onChange(newValueArray)
-      }
-
-      this.setState({dragCurrentIndex: null, dragOriginalIndex: null})
-    }
 
     this.state = {
-      currentY: null,
-      dragCurrentIndex: null,
-      dragOriginalIndex: null,
-      focusIndex: null,
       hasFocus: false
-    }
-  }
-
-  componentDidUpdate() {
-    if (this.cursor) {
-      this.setCursor(this.cursor)
-      this.cursor = null
     }
   }
 
@@ -211,206 +138,19 @@ export default class FieldStringEdit extends React.Component {
     return selectedOptions[0]
   }
 
-  handleDeleteRow(index) {
-    const {onChange, value: oldValueArray} = this.props
-    const newValueArray = oldValueArray.slice()
-
-    if (typeof onChange !== 'function') return
-
-    newValueArray.splice(index, 1)
-    onChange(newValueArray)
-  }
-
-  handleDragStart(index, {clientY}) {
-    const container = this.listContainerRef.getBoundingClientRect()
-    const draggedItem = this.itemRefs[index].getBoundingClientRect()
-
-    this.dragBottomLimit = container.height - draggedItem.height
-    this.dragStart = {
-      clientY,
-      top: draggedItem.top - container.top,
-      left: draggedItem.left - container.left,
-      height: draggedItem.height,
-      width: draggedItem.width
-    }
+  handleFocusChange(hasFocus) {
     this.setState({
-      dragOriginalIndex: index,
-      dragCurrentIndex: index,
-      currentY: clientY
-    })
-
-    window.document.addEventListener('mousemove', this.mousePositionListener)
-    window.document.addEventListener('mouseup', this.dragEndListener)
-  }
-
-  handleFocusChange(hasFocus, focusIndex) {
-    this.setState({
-      hasFocus,
-      focusIndex
+      hasFocus
     })
   }
 
   handleInputChange(value, options) {
-    const {
-      onChange,
-      schema: {format, publish = {}}
-    } = this.props
+    const {onChange} = this.props
 
-    if (typeof onChange !== 'function') return
-
-    // We prefer sending a `null` over an empty string.
-    onChange(value || null, options)
-
-    if (!publish.options && !format) {
-      const {selectionStart, selectionEnd} = this.inputRefs[0]
-
-      this.cursor = {selectionRange: [selectionStart, selectionEnd]}
+    if (typeof onChange === 'function') {
+      // We prefer sending a `null` over an empty string.
+      onChange(value || null, options)
     }
-  }
-
-  handleKeyDown(index, e) {
-    const {
-      onChange,
-      schema: {publish = {}},
-      value: valueArray
-    } = this.props
-    const {value, selectionStart, selectionEnd} = this.inputRefs[index]
-
-    switch (e.key) {
-      case 'Enter':
-        if (
-          // For multiline strings, this functionality is under Ctrl+Enter.
-          (!publish.multiline || e.ctrlKey) &&
-          Array.isArray(valueArray) &&
-          index < valueArray.length &&
-          typeof onChange === 'function'
-        ) {
-          // Add line break at cursor position, splitting the item in two.
-          const updatedRowValue = value.slice(0, selectionStart)
-          const addedRowValue = value.slice(selectionEnd)
-          const newValueArray = [
-            ...valueArray.slice(0, index),
-            updatedRowValue,
-            addedRowValue,
-            ...valueArray.slice(index + 1)
-          ]
-
-          onChange(newValueArray)
-
-          this.cursor = {index: index + 1, selectionRange: [0, 0]}
-        }
-
-        return
-
-      case 'Backspace':
-        if (
-          selectionStart === 0 &&
-          selectionEnd === 0 &&
-          index < valueArray.length &&
-          index > 0 &&
-          typeof onChange === 'function'
-        ) {
-          // Necessary – otherwise removes last char of previous item.
-          e.preventDefault()
-
-          // Concatenate the item with the previous one.
-          const prevItem = valueArray[index - 1] || ''
-          const currItem = valueArray[index] || ''
-          const newValueArray = [
-            ...valueArray.slice(0, index - 1),
-            prevItem + currItem,
-            ...valueArray.slice(index + 1)
-          ]
-
-          onChange(newValueArray)
-
-          this.cursor = {
-            index: index - 1,
-            selectionRange: [prevItem.length, prevItem.length]
-          }
-        } else if (index === valueArray.length) {
-          e.preventDefault()
-
-          this.setCursor({index: index - 1})
-        }
-
-        return
-
-      case 'Delete':
-        if (
-          selectionStart === value.length &&
-          selectionEnd === value.length &&
-          index < valueArray.length - 1
-        ) {
-          // Necessary – otherwise removes first char of next item.
-          e.preventDefault()
-
-          // Concatenate the item with the next one.
-          const currItem = valueArray[index] || ''
-          const nextItem = valueArray[index + 1] || ''
-          const newValueArray = [
-            ...valueArray.slice(0, index),
-            currItem + nextItem,
-            ...valueArray.slice(index + 2)
-          ]
-
-          onChange(newValueArray)
-
-          this.cursor = {
-            index,
-            selectionRange: [currItem.length, currItem.length]
-          }
-        }
-
-        return
-
-      case 'ArrowUp': {
-        if (index > 0 && selectionStart === 0 && selectionEnd === 0) {
-          e.preventDefault()
-
-          this.setCursor({
-            index: index - 1,
-            selectionRange: [0, 0]
-          })
-        }
-
-        return
-      }
-
-      case 'ArrowDown': {
-        if (
-          index < valueArray.length &&
-          selectionStart === value.length &&
-          selectionEnd === value.length
-        ) {
-          e.preventDefault()
-
-          this.setCursor({
-            index: index + 1,
-            selectionRange: [0, 0]
-          })
-        }
-
-        return
-      }
-    }
-  }
-
-  handleListInputChange(index, e) {
-    const {onChange, value: oldValueArray} = this.props
-
-    if (typeof onChange !== 'function') return
-
-    const newValueArray = Array.isArray(oldValueArray)
-      ? oldValueArray.slice()
-      : []
-
-    newValueArray.splice(index, 1, e.target.value)
-    onChange(newValueArray)
-
-    const {selectionStart, selectionEnd} = this.inputRefs[index]
-
-    this.cursor = {index, selectionRange: [selectionStart, selectionEnd]}
   }
 
   render() {
@@ -526,7 +266,7 @@ export default class FieldStringEdit extends React.Component {
     const type = multiline ? 'multiline' : 'text'
     const readOnly = readonly === true
 
-    const link = getLink(value, display.link)
+    const link = formatLink(value, display.link)
 
     return (
       <Label
@@ -581,108 +321,11 @@ export default class FieldStringEdit extends React.Component {
       name,
       placeholder,
       required,
-      schema: {publish: publishBlock = {}},
+      schema: {publish: publishSettings = {}},
       value
     } = this.props
-    const {
-      currentY,
-      dragCurrentIndex,
-      dragOriginalIndex,
-      focusIndex,
-      hasFocus
-    } = this.state
-    const {dragStart, dragBottomLimit} = this
-    const {display = {}, multiline, readonly, resizable, rows} = publishBlock
-    const type = multiline ? 'multiline' : 'text'
-    const readOnly = readonly === true
-
-    const top = dragStart.top + currentY - dragStart.clientY
-    const draggedStyle = {
-      top: `${Math.min(dragBottomLimit, Math.max(0, top))}px`,
-      left: `${dragStart.left}px`,
-      height: `${dragStart.height}px`,
-      width: `${dragStart.width}px`
-    }
-
-    const renderItems =
-      value &&
-      value.map((itemValue, index) => {
-        const link = getLink(itemValue, display.link)
-        const itemStyles = new Style(styles, 'list-item')
-          .addIf('dragged', dragOriginalIndex === index)
-          .addIf('focused', focusIndex === index)
-
-        return (
-          <div
-            className={itemStyles.getClasses()}
-            key={index}
-            onBlur={() => this.setState({focusIndex: null})}
-            onFocus={() => this.setState({focusIndex: index})}
-            ref={ref => {
-              this.itemRefs[index] = ref
-            }}
-            style={dragOriginalIndex === index ? draggedStyle : {}}
-          >
-            <span
-              className={styles['icon-drag']}
-              onMouseDown={this.handleDragStart.bind(this, index)}
-            >
-              Drag
-            </span>
-
-            <TextInput
-              heightType="content"
-              inputRef={ref => {
-                this.inputRefs[index] = ref
-              }}
-              name={name}
-              onBlur={this.handleFocusChange.bind(this, false)}
-              onFocus={this.handleFocusChange.bind(this, true)}
-              onInput={this.handleListInputChange.bind(this, index)}
-              onKeyDown={this.handleKeyDown.bind(this, index)}
-              placeholder={placeholder}
-              readOnly={readOnly}
-              resizable={resizable}
-              rows={rows}
-              type={type}
-              value={itemValue}
-            />
-
-            {link && (
-              <Button
-                accent="neutral"
-                className={styles['link-preview']}
-                href={link}
-                openInNewWindow={true}
-                size="small"
-              >
-                Open in new window
-              </Button>
-            )}
-
-            <Button
-              accent="destruct"
-              className={styles['remove-button']}
-              onClick={this.handleDeleteRow.bind(this, index)}
-              size="small"
-            >
-              Remove
-            </Button>
-          </div>
-        )
-      })
-
-    if (renderItems && dragCurrentIndex !== null) {
-      renderItems.splice(
-        dragCurrentIndex + Number(dragCurrentIndex > dragOriginalIndex),
-        0,
-        <div
-          className={styles['drag-placeholder']}
-          key="placeholder"
-          style={{height: dragStart.height}}
-        />
-      )
-    }
+    const {hasFocus} = this.state
+    const readOnly = publishSettings.readonly === true
 
     return (
       <Label
@@ -697,36 +340,14 @@ export default class FieldStringEdit extends React.Component {
         hasFocus={hasFocus}
         label={displayName}
       >
-        {value && (
-          <div
-            className={styles['list-container']}
-            ref={ref => {
-              this.listContainerRef = ref
-            }}
-          >
-            {renderItems}
-          </div>
-        )}
-        <div className={styles['list-item']}>
-          <span className={styles['icon-add']}>Add</span>
-          <TextInput
-            heightType="content"
-            inputRef={ref => {
-              this.inputRefs[value.length] = ref
-            }}
-            name={name}
-            onBlur={this.handleFocusChange.bind(this, false)}
-            onFocus={this.handleFocusChange.bind(this, true, value.length)}
-            onInput={this.handleListInputChange.bind(this, value.length)}
-            onKeyDown={this.handleKeyDown.bind(this, value.length)}
-            placeholder={placeholder}
-            readOnly={readOnly}
-            resizable={resizable}
-            rows={rows}
-            type={type}
-            value={null}
-          />
-        </div>
+        <StringArray
+          name={name}
+          onChange={this.handleInputChange.bind(this)}
+          onFocus={this.handleFocusChange.bind(this)}
+          placeholder={placeholder}
+          publishSettings={publishSettings}
+          valuesArray={value}
+        />
       </Label>
     )
   }
@@ -779,11 +400,5 @@ export default class FieldStringEdit extends React.Component {
         input.options[i].selected = true
       }
     }
-  }
-
-  setCursor({index, selectionRange} = {}) {
-    typeof index === 'number' && this.inputRefs[index].focus()
-    selectionRange &&
-      this.inputRefs[index || 0].setSelectionRange(...selectionRange)
   }
 }
